+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.14.0 (2024-07-29)
+
+- **[Breaking]** Adjust Appsignal middleware to use `Appsignal.monitor`.  
+  To use the middleware the `appsignal` gem in version `>= 3.11.0` is required.  
+  The configuration of the middleware changed and now only requires one option `class_name` and an optional `namespace`.
+
 ## 0.13.0 (2023-11-07)
 
 - Allow adding multiple routing keys to the consumer configuration, configure method within consumer will only accept `routing_keys` array instead of `routing_key` string